Calculate Sample Size for Desired CV

Estimates the sample size needed to achieve a desired coefficient of variation for a given dilution scheme.

estimate_sample_size(target_cv, mu, sd, a, b, f, u, USL, n_sim = 10000, ...)

Arguments

target_cv

Desired coefficient of variation.

mu

Log-scale mean.

sd

Log-scale standard deviation.

a

Lower truncation bound.

b

Upper truncation bound.

f

Dilution factor.

u

Amount placed on plate.

USL

Upper specification limit.

n_sim

Number of simulations.

...

Additional arguments passed to cv_heterogeneous.

Value

Sample size needed.

Examples

if (FALSE) { # \dontrun{
estimate_sample_size(target_cv = 0.1, mu = 2, sd = 0.2,
                     a = 0, b = 300, f = 0.01, u = 0.1,
                     USL = 1000, n_sim = 10000)
} # }