Probability of detection estimation when diluted sample collected from a homogeneous batch.

prob_detection_homogeneous provides a probability of detection in the original sample when samples collected from a homogeneous batch.

prob_detection_homogeneous(lambda, a, b, f, u, USL, n_sim)

Arguments

lambda

the expected cell count (\(\lambda\)).

a

lower domain of the number of cell counts.

b

upper domain of the number of cell counts.

f

final dilution factor.

u

amount put on the plate.

USL

upper specification limit.

n_sim

number of simulations (large simulations provide a more precise estimation).

Value

Probability of detection when the diluted sample collected from a homogeneous batch.

Details

prob_detection_homogeneous provides a probability of detection when the diluted sample has homogeneous contaminants. We define the random variable \(X_{i}\) is the number of colonies on the \(i^{th}\) plate. In practice, the acceptance for countable numbers of colonies on a plate must be between 30 and 300. Therefore, we can utilise bounded distributions to model the number of colonies on a plate. In the homogeneous case, we employed truncated Poisson distribution to model (this section will be updated later on).

Examples

lambda <- 2000
a <- 0
b <- 300
f <- 0.01
u <- 0.1
USL <- 1000
n_sim <- 50000
prob_detection_homogeneous(lambda, a, b, f, u, USL, n_sim)
#> [1] 0.6871